Move createCompactionJobs to AbstractCompactionStrategy
