+++ java/compaction-job-creation/src/main/java/sleeper/compaction/strategy/impl/SizeRatioCompactionStrategy.java
@@ -15,19 +15,6 @@
  */
 package sleeper.compaction.strategy.impl;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import sleeper.compaction.job.CompactionJob;
-import sleeper.core.partition.Partition;
-import sleeper.statestore.FileInfo;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
-
 /**
  * A {@link sleeper.compaction.strategy.CompactionStrategy} that is similar to the strategy used in Accumulo.
  * Given files sorted into increasing size order, it tests whether the sum of the
@@ -47,53 +34,8 @@ import java.util.stream.Collectors;
  * controls how many jobs can be running concurrently for each partition.
  */
 public class SizeRatioCompactionStrategy extends AbstractCompactionStrategy {
-    private static final Logger LOGGER = LoggerFactory.getLogger(SizeRatioCompactionStrategy.class);
 
     public SizeRatioCompactionStrategy() {
         super(new SizeRatioLeafStrategy(), new ShouldCreateJobsWithMaxPerPartition());
     }
-
-    @Override
-    public List<CompactionJob> createCompactionJobs(List<FileInfo> activeFilesWithJobId, List<FileInfo> activeFilesWithNoJobId, List<Partition> allPartitions) {
-        // Get list of partition ids from the above files
-        Set<String> partitionIds = activeFilesWithNoJobId.stream()
-                .map(FileInfo::getPartitionId)
-                .collect(Collectors.toSet());
-
-        // Get map from partition id to partition
-        Map<String, Partition> partitionIdToPartition = new HashMap<>();
-        for (Partition partition : allPartitions) {
-            partitionIdToPartition.put(partition.getId(), partition);
-        }
-
-        // Loop through partitions for the active files with no job id
-        List<CompactionJob> compactionJobs = new ArrayList<>();
-        for (String partitionId : partitionIds) {
-            Partition partition = partitionIdToPartition.get(partitionId);
-            if (null == partition) {
-                throw new RuntimeException("Cannot find partition for partition id " + partitionId);
-            }
-
-            if (partition.isLeafPartition()) {
-                long maxNumberOfJobsToCreate = shouldCreateJobsStrategy.maxCompactionJobsToCreate(
-                        partition, activeFilesWithJobId, activeFilesWithNoJobId);
-                if (maxNumberOfJobsToCreate < 1) {
-                    continue;
-                }
-                LOGGER.info("Max jobs to create = {}", maxNumberOfJobsToCreate);
-                List<CompactionJob> jobs = leafStrategy.createJobsForLeafPartition(partition, activeFilesWithNoJobId);
-                LOGGER.info("Defined {} compaction job{} for partition {}", jobs.size(), 1 == jobs.size() ? "s" : "", partitionId);
-                while (jobs.size() > maxNumberOfJobsToCreate) {
-                    jobs.remove(jobs.size() - 1);
-                }
-                LOGGER.info("Created {} compaction job{} for partition {}", jobs.size(), 1 == jobs.size() ? "s" : "", partitionId);
-                compactionJobs.addAll(jobs);
-            } else {
-                compactionJobs.addAll(createJobsForNonLeafPartition(partition, activeFilesWithNoJobId, partitionIdToPartition));
-            }
-        }
-
-        return compactionJobs;
-    }
-
 }
